Clarify error messages when reading files
Previously, trying to operate on a non-existing data or signature file
only printed a generic IO error, without mentioning the operation nor
the actual expected file. This was particularly unintuitive as I did not
know whether the tool expected to be pointed to the file or to the
signature in the first place, and it took a few attempts to figure it
out (or perhaps I should have just read the documentation...).